refactor: replaces location of actions from reducer to acrions file
