+++ src/client/src/index.js
@@ -7,9 +7,8 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import "monday-ui-react-core/dist/main.css";
-import { ItemsReducers } from './reducers/items-entities-reducer';
-
-store.dispatch(ItemsReducers.GetTodos());
+import {GetTodos} from './actions/ItemsActions';
+store.dispatch(GetTodos());
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
